refactor: Add encoding.go to handle JSON encoding
